refactor: new folder structure for components
* a bit easier
