+++ components/content/index.ts
@@ -1,0 +1,4 @@
+export { default as Text } from '../content/Text';
+export { default as Card } from './Card';
+export { default as DataTable } from './DataTable';
+export { default as Heading } from './Heading';
