--- components/datadisplay/index.ts
@@ -1,2 +1,0 @@
-export { default as Card } from './Card';
-export { default as DataTable } from './DataTable';
