+++ components/layout/index.ts
@@ -1,4 +1,6 @@
 export { default as Container } from './Container';
+export { default as DarkModeToggle } from './DarkModeToggle';
 export { default as Footer } from './Footer';
 export { default as Header } from './Header';
 export { default as Layout } from './Layout';
+export { default as UserMenu } from './UserMenu';
