--- components/navigation/index.ts
@@ -1,1 +1,0 @@
-export { default as UserMenu } from './UserMenu';
