+++ pages/admin/events/[id].tsx
@@ -17,8 +17,7 @@ import { Button } from 'components/inputs';
 import { notifications } from '@mantine/notifications';
 import { Container, Layout } from 'components/layout';
 import { Loading, Unauthorized } from 'components/feedback';
-import { Heading } from 'components/typography';
-import { DataTable } from 'components/datadisplay';
+import { DataTable, Heading } from 'components/content';
 import { EmailDrawer, RegistrationDrawer } from 'components/overlays';
 import { createColumnHelper } from '@tanstack/react-table';
 import Link from 'next/link';
